Feat: 검색 조건 유지하면서 페이징 처리 완료

+++ src/main/java/techit/gongsimchae/domain/portion/subdivision/repository/SubdivisionCustomRepositoryImpl.java
@@ -138,7 +138,8 @@ public class SubdivisionCustomRepositoryImpl implements SubdivisionCustomReposit
                 .where(subdivision.deleteStatus.eq(false).and(IsSale(searchDto)))
                 .fetchOne();
 
-        return new PageImpl<>(results, pageable,size);
+
+        return new PageImpl<>(results, pageable,size.intValue());
 
     }
